fix bugs not yet completed

+++ 3-semestre/estruturas-de-dados/arvores/avl/avl.h
@@ -14,7 +14,7 @@ avl_node_t avl_node_new(int value);
 void avl_node_insert(avl_node_t *node, int value);
 size_t avl_node_height(avl_node_t node);
 void avl_node_set_height(avl_node_t node);
-size_t avl_node_balance(avl_node_t node);
+long avl_node_balance(avl_node_t node);
 
 void avl_node_rot(avl_node_t *node);
 void avl_node_rot_ll(avl_node_t *node);
